Для DepositionRequest paymentParams.pof_offerAccepted обязателен
