+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v2.3.0 от 08.04.2021
+* Для DepositionRequest paymentParams.pof_offerAccepted обязателен
+
 ### v2.2.0 от 01.04.2021
 * Добавлен параметр -noverify для расшифровки ответа кассы
 * Генерация CSR со старым закрытым ключом
